Name Refactor
Some explanation:
"Surface" holds the back-end stuff, Math, Rendering, Input, Etc.
"Sandbox" will be the executable which people use to build games and compile them.

Sure, developers can build the game in raw code, but making an Editor which uses the Surface Engine itself would be great example material for developers who want to build it "code-only" style, create their own version of the Sandbox Editor itself, and also for those who just want to use an editor.

Naming convention:
The base class header file which changes implementation depending on platform has it's definition in the "Surface" folder, with a similarly named folder that contains the different implementations.

For GLFW, it's implemented in respective "GL{class}" objects. In the future with DirectX and Vulkan, they could have "DX{class}" and "VL{class}" names.